requirements.txt
requests
dnspython
scapy